[refactor] Put things related to page trailer into page0zip_trailer and things related to log to page0zip_mlog
Summary:
Added two new files, page0zip_trailer contains functions that writes to page trailer,
page0zip_mlog contains functions that writes to modification log.

Test Plan: jenkins

Reviewers: over

Reviewed By: over
