+++ storage/innobase/include/page0zip.ic
@@ -32,6 +32,7 @@ Created June 2005 by Marko Makela
 #include "page0zip.h"
 #include "mtr0log.h"
 #include "page0page.h"
+#include "page0zip_mlog.h"
 
 /* The format of compressed pages is as follows.
 
@@ -364,16 +365,6 @@ page_zip_des_init(
 }
 
 /**********************************************************************//**
-Write a log record of writing to the uncompressed header portion of a page. */
-UNIV_INTERN
-void
-page_zip_write_header_log(
-/*======================*/
-	const byte*	data,/*!< in: data on the uncompressed page */
-	ulint		length,	/*!< in: length of the data */
-	mtr_t*		mtr);	/*!< in: mini-transaction */
-
-/**********************************************************************//**
 Write data to the uncompressed header portion of a page.  The data must
 already have been written to the uncompressed page.
 However, the data portion of the uncompressed page may differ from
@@ -469,39 +460,6 @@ page_zip_compress_write_log_no_data(
 		mach_write_to_1(log_ptr, compression_flags);
 		mlog_close(mtr, log_ptr + 1);
 	}
-}
-
-/**********************************************************************//**
-Parses a log record of compressing an index page without the data.
-@return	end of log record or NULL */
-UNIV_INLINE
-byte*
-page_zip_parse_compress_no_data(
-/*============================*/
-	byte*		ptr,		/*!< in: buffer */
-	byte*		end_ptr,	/*!< in: buffer end */
-	page_t*		page,		/*!< in: uncompressed page */
-	page_zip_des_t*	page_zip,	/*!< out: compressed page */
-	dict_index_t*	index)		/*!< in: index */
-{
-	uchar	compression_flags;
-	if (end_ptr == ptr) {
-		return(NULL);
-	}
-
-	compression_flags = mach_read_from_1(ptr);
-
-	/* If page compression fails then there must be something wrong
-	because a compress log record is logged only if the compression
-	was successful. Crash in this case. */
-
-	if (page
-	    && !page_zip_compress(page_zip, page, index, compression_flags,
-				  NULL)) {
-		ut_error;
-	}
-
-	return(ptr + 1);
 }
 
 /**********************************************************************//**
